requirements.txt
aiohttp==3.7.3
aiosqlite==0.16.0
APScheduler==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
discord-ext-menus==1.0.0a27+g6f2b873
discord.py==1.6.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
PyNaCl==1.3.0
pytz==2020.5
requests==2.25.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
wrapt==1.12.1
yarl==1.5.1